# Faster `AlignedBuffer`