--- a/crates/contracts/core/ab-contracts-executor/src/lib.rs
+++ b/crates/contracts/core/ab-contracts-executor/src/lib.rs
@@ -1,4 +1,5 @@
 #![feature(
+    box_vec_non_null,
     get_mut_unchecked,
     pointer_is_aligned_to,
     ptr_as_ref_unchecked,