--- a/crates/contracts/core/ab-contracts-test-utils/src/lib.rs
+++ b/crates/contracts/core/ab-contracts-test-utils/src/lib.rs
@@ -1,5 +1,4 @@
 #![no_std]
-extern crate alloc;
 
 pub mod dummy_wallet;
 pub mod transaction_builder;