This replaced `Arc` with a more efficient and internally simpler implementation, also `LazyLock` usage was replaced with a static unconditionally available instance.

As the result performance improved even further:
```
flipper/direct          time:   [68.005 ns 68.284 ns 68.600 ns]
                        thrpt:  [14.577 Melem/s 14.645 Melem/s 14.705 Melem/s]
flipper/transaction     time:   [71.065 ns 71.352 ns 71.711 ns]
                        thrpt:  [13.945 Melem/s 14.015 Melem/s 14.072 Melem/s]

example-wallet/execute-only
                        time:   [3.0973 µs 3.1050 µs 3.1141 µs]
                        thrpt:  [321.12 Kelem/s 322.07 Kelem/s 322.86 Kelem/s]
```

The only thing preventing `NativeExecutor` from being usable in `no_std` environment with `alloc` is `HashMap` usage, which can be replaced with `BTreeMap`, but will be slower. Once there is a need for that, a new feature can be added to make it work in `no_std` environment.